// shared/enums.ts
// This file was generated by a custom prisma generator, do not edit manually.

export const InstitutionRole = {
  Admin: 'Admin',
  Educator: 'Educator',
  Member: 'Member',
} as const

// eslint-disable-next-line @typescript-eslint/no-redeclare -- intentionally naming the variable the same as the type
export type InstitutionRole = (typeof InstitutionRole)[keyof typeof InstitutionRole]

export const GroupRole = {
  Owner: 'Owner',
  Educator: 'Educator',
  Member: 'Member',
} as const

// eslint-disable-next-line @typescript-eslint/no-redeclare -- intentionally naming the variable the same as the type
export type GroupRole = (typeof GroupRole)[keyof typeof GroupRole]

export const QuestionType = {
  File: 'File',
  Text: 'Text',
  Options: 'Options',
} as const

// eslint-disable-next-line @typescript-eslint/no-redeclare -- intentionally naming the variable the same as the type
export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType]